+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v4.1.1]
 
 - Fixed documentation tables
+- OAuth1 `authenticate` returns `tokenData`
 
 ## [v4.1.0]