+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed documentation tables
 - OAuth1 `authenticate` returns `tokenData`
+- Updated documentation on SSL pinning
 
 ## [v4.1.0]