+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed documentation tables
 - OAuth1 `authenticate` returns `tokenData`
 - Updated documentation on SSL pinning
+- Add `logout` function to `sdk.auth`
 
 ## [v4.1.0]