+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - OAuth1 `authenticate` returns `tokenData`
 - Updated documentation on SSL pinning
 - Add `logout` function to `sdk.auth`
+- Updated installation instructions
 
 ## [v4.1.0]